pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.8",
    "ipykernel>=7.2.0",
    "ipython>=9.12.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.15",
    "langgraph>=1.1.8",
    "langgraph-cli[inmem]>=0.4.24",
    "python-dotenv>=1.2.2",
]
